+++ baybe/searchspace/core.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import warnings
 from collections.abc import Iterable, Sequence
 from enum import Enum
 from typing import cast
@@ -303,23 +304,76 @@ class SearchSpace(SerialMixin):
 
     def transform(
         self,
-        data: pd.DataFrame,
+        df: pd.DataFrame | None = None,
+        /,
+        *,
+        allow_missing: bool = False,
+        allow_extra: bool | None = None,
+        data: pd.DataFrame | None = None,
     ) -> pd.DataFrame:
-        """Transform data from experimental to computational representation.
-
-        This function can e.g. be used to transform data obtained from measurements.
-        Continuous parameters are not transformed but included.
+        """Transform parameters from experimental to computational representation.
 
         Args:
-            data: The data to be transformed. Must contain all specified parameters, can
-                contain more columns.
+            df: The dataframe to be transformed. The allowed columns of the dataframe
+                are dictated by the ``allow_missing`` and ``allow_extra`` flags.
+                The ``None`` default value is for temporary backward compatibility only
+                and will be removed in a future version.
+            allow_missing: If ``False``, each parameter of the space must have
+                (exactly) one corresponding column in the given dataframe. If ``True``,
+                the dataframe may contain only a subset of parameter columns.
+            allow_extra: If ``False``, every column present in the dataframe must
+                correspond to (exactly) one parameter of the space. If ``True``, the
+                dataframe may contain additional non-parameter-related columns, which
+                will be ignored.
+                The ``None`` default value is for temporary backward compatibility only
+                and will be removed in a future version.
+            data: Ignore! For backward compatibility only.
+
+        Raises:
+            ValueError: If dataframes are passed to both ``df`` and ``data``.
 
         Returns:
-            A dataframe with the parameters in computational representation.
+            A corresponding dataframe with parameters in computational representation.
         """
+        # TODO: Remove deprecation-related explanation of `None` default values
+        #   from docstring once deprecation expires
+        # >>>>>>>>>> Deprecation
+        if not ((df is None) ^ (data is None)):
+            raise ValueError(
+                "Provide the dataframe to be transformed as argument to `df`."
+            )
+
+        if data is not None:
+            df = data
+            warnings.warn(
+                "Providing the dataframe via the `data` argument is deprecated and "
+                "will be removed in a future version. Please pass your dataframe "
+                "as positional argument instead.",
+                DeprecationWarning,
+            )
+
+        # Mypy does not infer from the above that `df` must be a dataframe here
+        assert isinstance(df, pd.DataFrame)
+
+        if allow_extra is None:
+            allow_extra = True
+            if set(df) - {p.name for p in self.parameters}:
+                warnings.warn(
+                    "For backward compatibility, the new `allow_extra` flag is set "
+                    "to `True` when left unspecified. However, this behavior will be "
+                    "changed in a future version. If you want to invoke the old "
+                    "behavior, please explicitly set `allow_extra=True`.",
+                    DeprecationWarning,
+                )
+        # <<<<<<<<<< Deprecation
+
         # Transform subspaces separately
-        df_discrete = self.discrete.transform(data)
-        df_continuous = self.continuous.transform(data)
+        df_discrete = self.discrete.transform(
+            df, allow_missing=allow_missing, allow_extra=allow_extra
+        )
+        df_continuous = self.continuous.transform(
+            df, allow_missing=allow_missing, allow_extra=allow_extra
+        )
 
         # Combine Subspaces
         comp_rep = pd.concat([df_discrete, df_continuous], axis=1)